Rename private submodule to _parallel_reductions
