+++ sklearn/metrics/pairwise.py
@@ -31,7 +31,7 @@ from ..utils._mask import _get_mask
 from ..utils.fixes import delayed
 from ..utils.fixes import sp_version, parse_version
 
-from ._argkmin_fast import FastSquaredEuclideanArgKmin
+from ._parallel_reductions import FastSquaredEuclideanArgKmin
 from ._pairwise_fast import _chi2_kernel_fast, _sparse_manhattan
 from ..exceptions import DataConversionWarning
